+++ android/src/main/java/com/twiliovoicereactnative/Constants.java
@@ -2,6 +2,7 @@ package com.twiliovoicereactnative;
 
 public class Constants {
   public static final String CALL_SID_KEY = "CALL_SID";
+  public static final String VOICE_CHANNEL_GROUP = "notification-group";
   public static final String VOICE_CHANNEL_LOW_IMPORTANCE = "notification-channel-low-importance";
   public static final String VOICE_CHANNEL_HIGH_IMPORTANCE = "notification-channel-high-importance";
   public static final String VOICE_CHANNEL_DEFAULT_IMPORTANCE = "notification-channel-normal-importance";